Move calculateMoveDates into a new file.
