+++ pkg/handlers/internalapi/moves.go
@@ -6,9 +6,6 @@ import (
 	"github.com/go-openapi/runtime/middleware"
 	"github.com/go-openapi/swag"
 	"github.com/gobuffalo/uuid"
-	"github.com/rickar/cal"
-	"github.com/transcom/mymove/pkg/route"
-	"github.com/transcom/mymove/pkg/unit"
 	"go.uber.org/zap"
 
 	"github.com/transcom/mymove/pkg/auth"
@@ -241,81 +238,4 @@ func (h ShowMoveDatesSummaryHandler) Handle(params moveop.ShowMoveDatesSummaryPa
 	}
 
 	return moveop.NewShowMoveDatesSummaryOK().WithPayload(moveDatesSummary)
-}
-
-// MoveDateSummary contains the set of dates for a move
-type MoveDateSummary struct {
-	PackDays     []time.Time
-	PickupDays   []time.Time
-	TransitDays  []time.Time
-	DeliveryDays []time.Time
-	ReportDays   []time.Time
-}
-
-func calculateMoveDates(move models.Move, planner route.Planner, moveDate time.Time) (MoveDateSummary, error) {
-	summary := MoveDateSummary{}
-
-	transitDistance, err := planner.TransitDistance(&move.Orders.ServiceMember.DutyStation.Address,
-		&move.Orders.NewDutyStation.Address)
-	if err != nil {
-		return summary, err
-	}
-
-	entitlementWeight := unit.Pound(models.GetEntitlement(*move.Orders.ServiceMember.Rank, move.Orders.HasDependents,
-		move.Orders.SpouseHasProGear))
-
-	numTransitDays, err := models.TransitDays(entitlementWeight, transitDistance)
-	if err != nil {
-		return summary, err
-	}
-
-	numPackDays := models.PackDays(entitlementWeight)
-	usCalendar := handlers.NewUSCalendar()
-
-	lastPossiblePackDay := moveDate.AddDate(0, 0, -1)
-	summary.PackDays = createPastMoveDates(lastPossiblePackDay, numPackDays, false, usCalendar)
-
-	firstPossiblePickupDay := moveDate
-	pickupDays := createFutureMoveDates(firstPossiblePickupDay, 1, false, usCalendar)
-	summary.PickupDays = pickupDays
-
-	firstPossibleTransitDay := time.Time(pickupDays[len(pickupDays)-1]).AddDate(0, 0, 1)
-	transitDays := createFutureMoveDates(firstPossibleTransitDay, numTransitDays, true, usCalendar)
-	summary.TransitDays = transitDays
-
-	firstPossibleDeliveryDay := time.Time(transitDays[len(transitDays)-1]).AddDate(0, 0, 1)
-	summary.DeliveryDays = createFutureMoveDates(firstPossibleDeliveryDay, 1, false, usCalendar)
-
-	summary.ReportDays = []time.Time{move.Orders.ReportByDate.UTC()}
-
-	return summary, nil
-}
-
-func createFutureMoveDates(startDate time.Time, numDays int, includeWeekendsAndHolidays bool, calendar *cal.Calendar) []time.Time {
-	dates := make([]time.Time, 0, numDays)
-
-	daysAdded := 0
-	for d := startDate; daysAdded < numDays; d = d.AddDate(0, 0, 1) {
-		if includeWeekendsAndHolidays || calendar.IsWorkday(d) {
-			dates = append(dates, d)
-			daysAdded++
-		}
-	}
-
-	return dates
-}
-
-func createPastMoveDates(startDate time.Time, numDays int, includeWeekendsAndHolidays bool, calendar *cal.Calendar) []time.Time {
-	dates := make([]time.Time, numDays)
-
-	daysAdded := 0
-	for d := startDate; daysAdded < numDays; d = d.AddDate(0, 0, -1) {
-		if includeWeekendsAndHolidays || calendar.IsWorkday(d) {
-			// Since we're working backwards, put dates at end of slice.
-			dates[numDays-daysAdded-1] = d
-			daysAdded++
-		}
-	}
-
-	return dates
 }
